changes made for the file structures, java oop objects , java docs
